// src/app/modules/mktg/interfaces/sro.interface.ts
export interface ISRO {
    _id?: string;
    umid?: string;
    fullName?: string;
    avg?: string;
    q111?: number;
    q112?: number;
    q113?: number;
    q121?: number;
    q122?: number;
    q123?: number;
    q13?: string;
    q141?: number;
    q142?: number;
    q151?: number;
    q152?: number;
    q153?: number;
    q161?: number;
    q162?: number;
    q163?: number;
    q171?: number;
    q211?: number;
    q212?: number;
    q213?: number;
    q221?: number;
    q222?: number;
    q223?: number;
    q231?: number;
    q232?: number;
    q233?: number;
    q241?: number;
    q242?: number;
    q243?: number;
    q251?: number;
    q252?: number;
    q253?: number;
    q261?: number;
    q262?: number;
    q263?: number;
    q271?: number;
    q272?: number;
    q273?: number;
    q281?: number;
    q282?: number;
    q283?: number;
    q284?: number;
    q285?: number;
    q286?: number;
    q287?: number;
    q288?: number;
    q29?: string;
    q31?: string;
    q32?: string;
    q33?: string;
    q34?: number;
    q351?: number;
    q352?: number;
    q353?: number;
    q354?: number;
    q41?: string;
    q42?: number;
    q43?: string;
    q44?: string;
    q45?: string;
    surveyAt?: Date | string;
    comment?: string;
    createdAt?: Date | string;
    updatedAt?: Date | string;
}